requirement.txt
python-libnmap
scapy-python3